# packages/shared/pkg/http/edge/api.gen.go
// Package api provides primitives to interact with the openapi HTTP API.
//
// Code generated by github.com/oapi-codegen/oapi-codegen/v2 version v2.4.1 DO NOT EDIT.
package api

import (
	"fmt"
	"net/http"

	"github.com/gin-gonic/gin"
	"github.com/oapi-codegen/runtime"
)

// ServerInterface represents all server handlers.
type ServerInterface interface {

	// (GET /health)
	HealthCheck(c *gin.Context)

	// (GET /health/machine)
	HealthCheckMachine(c *gin.Context)

	// (GET /health/traffic)
	HealthCheckTraffic(c *gin.Context)

	// (GET /v1/info)
	V1Info(c *gin.Context)
	// Get latest metrics for multiple sandboxes
	// (GET /v1/sandboxes/metrics)
	V1SandboxesMetrics(c *gin.Context, params V1SandboxesMetricsParams)
	// List structured sandbox logs
	// (GET /v1/sandboxes/{sandboxID}/logs)
	V1SandboxLogs(c *gin.Context, sandboxID string, params V1SandboxLogsParams)
	// Get time-series metrics for a sandbox
	// (GET /v1/sandboxes/{sandboxID}/metrics)
	V1SandboxMetrics(c *gin.Context, sandboxID string, params V1SandboxMetricsParams)

	// (GET /v1/service-discovery/nodes)
	V1ServiceDiscoveryNodes(c *gin.Context)

	// (POST /v1/service-discovery/nodes/drain)
	V1ServiceDiscoveryNodeDrain(c *gin.Context)

	// (POST /v1/service-discovery/nodes/kill)
	V1ServiceDiscoveryNodeKill(c *gin.Context)
	// Get the orchestrators
	// (GET /v1/service-discovery/nodes/orchestrators)
	V1ServiceDiscoveryGetOrchestrators(c *gin.Context)
	// Template build logs
	// (GET /v1/templates/builds/{buildID}/logs)
	V1TemplateBuildLogs(c *gin.Context, buildID string, params V1TemplateBuildLogsParams)
}

// ServerInterfaceWrapper converts contexts to parameters.
type ServerInterfaceWrapper struct {
	Handler            ServerInterface
	HandlerMiddlewares []MiddlewareFunc
	ErrorHandler       func(*gin.Context, error, int)
}

type MiddlewareFunc func(c *gin.Context)

// HealthCheck operation middleware
func (siw *ServerInterfaceWrapper) HealthCheck(c *gin.Context) {

	for _, middleware := range siw.HandlerMiddlewares {
		middleware(c)
		if c.IsAborted() {
			return
		}
	}

	siw.Handler.HealthCheck(c)
}

// HealthCheckMachine operation middleware
func (siw *ServerInterfaceWrapper) HealthCheckMachine(c *gin.Context) {

	for _, middleware := range siw.HandlerMiddlewares {
		middleware(c)
		if c.IsAborted() {
			return
		}
	}

	siw.Handler.HealthCheckMachine(c)
}

// HealthCheckTraffic operation middleware
func (siw *ServerInterfaceWrapper) HealthCheckTraffic(c *gin.Context) {

	for _, middleware := range siw.HandlerMiddlewares {
		middleware(c)
		if c.IsAborted() {
			return
		}
	}

	siw.Handler.HealthCheckTraffic(c)
}

// V1Info operation middleware
func (siw *ServerInterfaceWrapper) V1Info(c *gin.Context) {

	for _, middleware := range siw.HandlerMiddlewares {
		middleware(c)
		if c.IsAborted() {
			return
		}
	}

	siw.Handler.V1Info(c)
}

// V1SandboxesMetrics operation middleware
func (siw *ServerInterfaceWrapper) V1SandboxesMetrics(c *gin.Context) {

	var err error

	c.Set(ApiKeyAuthScopes, []string{})

	// Parameter object where we will unmarshal all parameters from the context
	var params V1SandboxesMetricsParams

	// ------------- Required query parameter "teamID" -------------

	if paramValue := c.Query("teamID"); paramValue != "" {

	} else {
		siw.ErrorHandler(c, fmt.Errorf("Query argument teamID is required, but not found"), http.StatusBadRequest)
		return
	}

	err = runtime.BindQueryParameter("form", true, true, "teamID", c.Request.URL.Query(), &params.TeamID)
	if err != nil {
		siw.ErrorHandler(c, fmt.Errorf("Invalid format for parameter teamID: %w", err), http.StatusBadRequest)
		return
	}

	// ------------- Required query parameter "sandbox_ids" -------------

	if paramValue := c.Query("sandbox_ids"); paramValue != "" {

	} else {
		siw.ErrorHandler(c, fmt.Errorf("Query argument sandbox_ids is required, but not found"), http.StatusBadRequest)
		return
	}

	err = runtime.BindQueryParameter("form", true, true, "sandbox_ids", c.Request.URL.Query(), &params.SandboxIds)
	if err != nil {
		siw.ErrorHandler(c, fmt.Errorf("Invalid format for parameter sandbox_ids: %w", err), http.StatusBadRequest)
		return
	}

	for _, middleware := range siw.HandlerMiddlewares {
		middleware(c)
		if c.IsAborted() {
			return
		}
	}

	siw.Handler.V1SandboxesMetrics(c, params)
}

// V1SandboxLogs operation middleware
func (siw *ServerInterfaceWrapper) V1SandboxLogs(c *gin.Context) {

	var err error

	// ------------- Path parameter "sandboxID" -------------
	var sandboxID string

	err = runtime.BindStyledParameterWithOptions("simple", "sandboxID", c.Param("sandboxID"), &sandboxID, runtime.BindStyledParameterOptions{Explode: false, Required: true})
	if err != nil {
		siw.ErrorHandler(c, fmt.Errorf("Invalid format for parameter sandboxID: %w", err), http.StatusBadRequest)
		return
	}

	c.Set(ApiKeyAuthScopes, []string{})

	// Parameter object where we will unmarshal all parameters from the context
	var params V1SandboxLogsParams

	// ------------- Required query parameter "teamID" -------------

	if paramValue := c.Query("teamID"); paramValue != "" {

	} else {
		siw.ErrorHandler(c, fmt.Errorf("Query argument teamID is required, but not found"), http.StatusBadRequest)
		return
	}

	err = runtime.BindQueryParameter("form", true, true, "teamID", c.Request.URL.Query(), &params.TeamID)
	if err != nil {
		siw.ErrorHandler(c, fmt.Errorf("Invalid format for parameter teamID: %w", err), http.StatusBadRequest)
		return
	}

	// ------------- Optional query parameter "cursor" -------------

	err = runtime.BindQueryParameter("form", true, false, "cursor", c.Request.URL.Query(), &params.Cursor)
	if err != nil {
		siw.ErrorHandler(c, fmt.Errorf("Invalid format for parameter cursor: %w", err), http.StatusBadRequest)
		return
	}

	// ------------- Optional query parameter "limit" -------------

	err = runtime.BindQueryParameter("form", true, false, "limit", c.Request.URL.Query(), &params.Limit)
	if err != nil {
		siw.ErrorHandler(c, fmt.Errorf("Invalid format for parameter limit: %w", err), http.StatusBadRequest)
		return
	}

	// ------------- Optional query parameter "direction" -------------

	err = runtime.BindQueryParameter("form", true, false, "direction", c.Request.URL.Query(), &params.Direction)
	if err != nil {
		siw.ErrorHandler(c, fmt.Errorf("Invalid format for parameter direction: %w", err), http.StatusBadRequest)
		return
	}

	// ------------- Optional query parameter "level" -------------

	err = runtime.BindQueryParameter("form", true, false, "level", c.Request.URL.Query(), &params.Level)
	if err != nil {
		siw.ErrorHandler(c, fmt.Errorf("Invalid format for parameter level: %w", err), http.StatusBadRequest)
		return
	}

	for _, middleware := range siw.HandlerMiddlewares {
		middleware(c)
		if c.IsAborted() {
			return
		}
	}

	siw.Handler.V1SandboxLogs(c, sandboxID, params)
}

// V1SandboxMetrics operation middleware
func (siw *ServerInterfaceWrapper) V1SandboxMetrics(c *gin.Context) {

	var err error

	// ------------- Path parameter "sandboxID" -------------
	var sandboxID string

	err = runtime.BindStyledParameterWithOptions("simple", "sandboxID", c.Param("sandboxID"), &sandboxID, runtime.BindStyledParameterOptions{Explode: false, Required: true})
	if err != nil {
		siw.ErrorHandler(c, fmt.Errorf("Invalid format for parameter sandboxID: %w", err), http.StatusBadRequest)
		return
	}

	c.Set(ApiKeyAuthScopes, []string{})

	// Parameter object where we will unmarshal all parameters from the context
	var params V1SandboxMetricsParams

	// ------------- Required query parameter "teamID" -------------

	if paramValue := c.Query("teamID"); paramValue != "" {

	} else {
		siw.ErrorHandler(c, fmt.Errorf("Query argument teamID is required, but not found"), http.StatusBadRequest)
		return
	}

	err = runtime.BindQueryParameter("form", true, true, "teamID", c.Request.URL.Query(), &params.TeamID)
	if err != nil {
		siw.ErrorHandler(c, fmt.Errorf("Invalid format for parameter teamID: %w", err), http.StatusBadRequest)
		return
	}

	// ------------- Optional query parameter "start" -------------

	err = runtime.BindQueryParameter("form", true, false, "start", c.Request.URL.Query(), &params.Start)
	if err != nil {
		siw.ErrorHandler(c, fmt.Errorf("Invalid format for parameter start: %w", err), http.StatusBadRequest)
		return
	}

	// ------------- Optional query parameter "end" -------------

	err = runtime.BindQueryParameter("form", true, false, "end", c.Request.URL.Query(), &params.End)
	if err != nil {
		siw.ErrorHandler(c, fmt.Errorf("Invalid format for parameter end: %w", err), http.StatusBadRequest)
		return
	}

	for _, middleware := range siw.HandlerMiddlewares {
		middleware(c)
		if c.IsAborted() {
			return
		}
	}

	siw.Handler.V1SandboxMetrics(c, sandboxID, params)
}

// V1ServiceDiscoveryNodes operation middleware
func (siw *ServerInterfaceWrapper) V1ServiceDiscoveryNodes(c *gin.Context) {

	c.Set(ApiKeyAuthScopes, []string{})

	for _, middleware := range siw.HandlerMiddlewares {
		middleware(c)
		if c.IsAborted() {
			return
		}
	}

	siw.Handler.V1ServiceDiscoveryNodes(c)
}

// V1ServiceDiscoveryNodeDrain operation middleware
func (siw *ServerInterfaceWrapper) V1ServiceDiscoveryNodeDrain(c *gin.Context) {

	c.Set(ApiKeyAuthScopes, []string{})

	for _, middleware := range siw.HandlerMiddlewares {
		middleware(c)
		if c.IsAborted() {
			return
		}
	}

	siw.Handler.V1ServiceDiscoveryNodeDrain(c)
}

// V1ServiceDiscoveryNodeKill operation middleware
func (siw *ServerInterfaceWrapper) V1ServiceDiscoveryNodeKill(c *gin.Context) {

	c.Set(ApiKeyAuthScopes, []string{})

	for _, middleware := range siw.HandlerMiddlewares {
		middleware(c)
		if c.IsAborted() {
			return
		}
	}

	siw.Handler.V1ServiceDiscoveryNodeKill(c)
}

// V1ServiceDiscoveryGetOrchestrators operation middleware
func (siw *ServerInterfaceWrapper) V1ServiceDiscoveryGetOrchestrators(c *gin.Context) {

	c.Set(ApiKeyAuthScopes, []string{})

	for _, middleware := range siw.HandlerMiddlewares {
		middleware(c)
		if c.IsAborted() {
			return
		}
	}

	siw.Handler.V1ServiceDiscoveryGetOrchestrators(c)
}

// V1TemplateBuildLogs operation middleware
func (siw *ServerInterfaceWrapper) V1TemplateBuildLogs(c *gin.Context) {

	var err error

	// ------------- Path parameter "buildID" -------------
	var buildID string

	err = runtime.BindStyledParameterWithOptions("simple", "buildID", c.Param("buildID"), &buildID, runtime.BindStyledParameterOptions{Explode: false, Required: true})
	if err != nil {
		siw.ErrorHandler(c, fmt.Errorf("Invalid format for parameter buildID: %w", err), http.StatusBadRequest)
		return
	}

	c.Set(ApiKeyAuthScopes, []string{})

	// Parameter object where we will unmarshal all parameters from the context
	var params V1TemplateBuildLogsParams

	// ------------- Optional query parameter "orchestratorID" -------------

	err = runtime.BindQueryParameter("form", true, false, "orchestratorID", c.Request.URL.Query(), &params.OrchestratorID)
	if err != nil {
		siw.ErrorHandler(c, fmt.Errorf("Invalid format for parameter orchestratorID: %w", err), http.StatusBadRequest)
		return
	}

	// ------------- Required query parameter "templateID" -------------

	if paramValue := c.Query("templateID"); paramValue != "" {

	} else {
		siw.ErrorHandler(c, fmt.Errorf("Query argument templateID is required, but not found"), http.StatusBadRequest)
		return
	}

	err = runtime.BindQueryParameter("form", true, true, "templateID", c.Request.URL.Query(), &params.TemplateID)
	if err != nil {
		siw.ErrorHandler(c, fmt.Errorf("Invalid format for parameter templateID: %w", err), http.StatusBadRequest)
		return
	}

	// ------------- Optional query parameter "offset" -------------

	err = runtime.BindQueryParameter("form", true, false, "offset", c.Request.URL.Query(), &params.Offset)
	if err != nil {
		siw.ErrorHandler(c, fmt.Errorf("Invalid format for parameter offset: %w", err), http.StatusBadRequest)
		return
	}

	// ------------- Optional query parameter "start" -------------

	err = runtime.BindQueryParameter("form", true, false, "start", c.Request.URL.Query(), &params.Start)
	if err != nil {
		siw.ErrorHandler(c, fmt.Errorf("Invalid format for parameter start: %w", err), http.StatusBadRequest)
		return
	}

	// ------------- Optional query parameter "end" -------------

	err = runtime.BindQueryParameter("form", true, false, "end", c.Request.URL.Query(), &params.End)
	if err != nil {
		siw.ErrorHandler(c, fmt.Errorf("Invalid format for parameter end: %w", err), http.StatusBadRequest)
		return
	}

	// ------------- Optional query parameter "limit" -------------

	err = runtime.BindQueryParameter("form", true, false, "limit", c.Request.URL.Query(), &params.Limit)
	if err != nil {
		siw.ErrorHandler(c, fmt.Errorf("Invalid format for parameter limit: %w", err), http.StatusBadRequest)
		return
	}

	// ------------- Optional query parameter "direction" -------------

	err = runtime.BindQueryParameter("form", true, false, "direction", c.Request.URL.Query(), &params.Direction)
	if err != nil {
		siw.ErrorHandler(c, fmt.Errorf("Invalid format for parameter direction: %w", err), http.StatusBadRequest)
		return
	}

	// ------------- Optional query parameter "level" -------------

	err = runtime.BindQueryParameter("form", true, false, "level", c.Request.URL.Query(), &params.Level)
	if err != nil {
		siw.ErrorHandler(c, fmt.Errorf("Invalid format for parameter level: %w", err), http.StatusBadRequest)
		return
	}

	for _, middleware := range siw.HandlerMiddlewares {
		middleware(c)
		if c.IsAborted() {
			return
		}
	}

	siw.Handler.V1TemplateBuildLogs(c, buildID, params)
}

// GinServerOptions provides options for the Gin server.
type GinServerOptions struct {
	BaseURL      string
	Middlewares  []MiddlewareFunc
	ErrorHandler func(*gin.Context, error, int)
}

// RegisterHandlers creates http.Handler with routing matching OpenAPI spec.
func RegisterHandlers(router gin.IRouter, si ServerInterface) {
	RegisterHandlersWithOptions(router, si, GinServerOptions{})
}

// RegisterHandlersWithOptions creates http.Handler with additional options
func RegisterHandlersWithOptions(router gin.IRouter, si ServerInterface, options GinServerOptions) {
	errorHandler := options.ErrorHandler
	if errorHandler == nil {
		errorHandler = func(c *gin.Context, err error, statusCode int) {
			c.JSON(statusCode, gin.H{"msg": err.Error()})
		}
	}

	wrapper := ServerInterfaceWrapper{
		Handler:            si,
		HandlerMiddlewares: options.Middlewares,
		ErrorHandler:       errorHandler,
	}

	router.GET(options.BaseURL+"/health", wrapper.HealthCheck)
	router.GET(options.BaseURL+"/health/machine", wrapper.HealthCheckMachine)
	router.GET(options.BaseURL+"/health/traffic", wrapper.HealthCheckTraffic)
	router.GET(options.BaseURL+"/v1/info", wrapper.V1Info)
	router.GET(options.BaseURL+"/v1/sandboxes/metrics", wrapper.V1SandboxesMetrics)
	router.GET(options.BaseURL+"/v1/sandboxes/:sandboxID/logs", wrapper.V1SandboxLogs)
	router.GET(options.BaseURL+"/v1/sandboxes/:sandboxID/metrics", wrapper.V1SandboxMetrics)
	router.GET(options.BaseURL+"/v1/service-discovery/nodes", wrapper.V1ServiceDiscoveryNodes)
	router.POST(options.BaseURL+"/v1/service-discovery/nodes/drain", wrapper.V1ServiceDiscoveryNodeDrain)
	router.POST(options.BaseURL+"/v1/service-discovery/nodes/kill", wrapper.V1ServiceDiscoveryNodeKill)
	router.GET(options.BaseURL+"/v1/service-discovery/nodes/orchestrators", wrapper.V1ServiceDiscoveryGetOrchestrators)
	router.GET(options.BaseURL+"/v1/templates/builds/:buildID/logs", wrapper.V1TemplateBuildLogs)
}